+++ packages/react-charts/src/components/ChartCursorContainer/ChartCursorContainer.tsx
@@ -239,7 +239,6 @@ export const ChartCursorContainer: React.FunctionComponent<ChartCursorContainerP
   );
 };
 ChartCursorContainer.displayName = 'ChartCursorContainer';
-ChartCursorContainer.defaultProps = (VictoryCursorContainer as any).defaultProps;
 
 // Note: VictoryCursorContainer.defaultEvents & VictoryContainer.role must be hoisted
 hoistNonReactStatics(ChartCursorContainer, VictoryCursorContainer);