+++ packages/react-integration/cypress/integration/toolbar.spec.ts
@@ -1,4 +1,4 @@
-xdescribe('Data Toolbar Demo Test', () => {
+describe('Data Toolbar Demo Test', () => {
   it('Navigate to demo section', () => {
     cy.visit('http://localhost:3000/toolbar-demo-nav-link');
   });