+++ packages/react-docs/CHANGELOG.md
@@ -3,6 +3,64 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.0.0-prerelease.25](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.0.0-prerelease.24...@patternfly/react-docs@7.0.0-prerelease.25) (2024-09-23)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+# [7.0.0-prerelease.24](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.0.0-prerelease.23...@patternfly/react-docs@7.0.0-prerelease.24) (2024-09-20)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+# [7.0.0-prerelease.23](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.0.0-prerelease.22...@patternfly/react-docs@7.0.0-prerelease.23) (2024-09-20)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+# [7.0.0-prerelease.22](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.0.0-prerelease.21...@patternfly/react-docs@7.0.0-prerelease.22) (2024-09-19)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+# [7.0.0-prerelease.21](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.0.0-prerelease.20...@patternfly/react-docs@7.0.0-prerelease.21) (2024-09-19)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+# [7.0.0-prerelease.20](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.0.0-prerelease.19...@patternfly/react-docs@7.0.0-prerelease.20) (2024-09-18)
+
+### Bug Fixes
+
+- **generate tokens:** prefix tokens with a t\_ ([#11002](https://github.com/patternfly/patternfly-react/issues/11002)) ([0dac6b8](https://github.com/patternfly/patternfly-react/commit/0dac6b88b9f560975627c80c6e8ab816d8b78ac9))
+
+# [7.0.0-prerelease.19](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.0.0-prerelease.18...@patternfly/react-docs@7.0.0-prerelease.19) (2024-09-18)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+# [7.0.0-prerelease.18](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.0.0-prerelease.17...@patternfly/react-docs@7.0.0-prerelease.18) (2024-09-18)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+# [7.0.0-prerelease.17](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.0.0-prerelease.16...@patternfly/react-docs@7.0.0-prerelease.17) (2024-09-17)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+# [7.0.0-prerelease.16](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.0.0-prerelease.15...@patternfly/react-docs@7.0.0-prerelease.16) (2024-09-16)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+# [7.0.0-prerelease.15](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.0.0-prerelease.14...@patternfly/react-docs@7.0.0-prerelease.15) (2024-09-16)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+# [7.0.0-prerelease.14](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.0.0-prerelease.13...@patternfly/react-docs@7.0.0-prerelease.14) (2024-09-13)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+# [7.0.0-prerelease.13](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.0.0-prerelease.12...@patternfly/react-docs@7.0.0-prerelease.13) (2024-09-13)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+# [7.0.0-prerelease.12](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.0.0-prerelease.11...@patternfly/react-docs@7.0.0-prerelease.12) (2024-09-12)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
 # [7.0.0-prerelease.11](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@7.0.0-prerelease.10...@patternfly/react-docs@7.0.0-prerelease.11) (2024-09-12)
 
 **Note:** Version bump only for package @patternfly/react-docs