+++ packages/react-templates/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.0.0-prerelease.15](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-templates@6.0.0-prerelease.14...@patternfly/react-templates@6.0.0-prerelease.15) (2024-09-23)
+
+**Note:** Version bump only for package @patternfly/react-templates
+
+# [6.0.0-prerelease.14](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-templates@6.0.0-prerelease.13...@patternfly/react-templates@6.0.0-prerelease.14) (2024-09-20)
+
+**Note:** Version bump only for package @patternfly/react-templates
+
+# [6.0.0-prerelease.13](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-templates@6.0.0-prerelease.12...@patternfly/react-templates@6.0.0-prerelease.13) (2024-09-19)
+
+**Note:** Version bump only for package @patternfly/react-templates
+
+# [6.0.0-prerelease.12](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-templates@6.0.0-prerelease.11...@patternfly/react-templates@6.0.0-prerelease.12) (2024-09-18)
+
+**Note:** Version bump only for package @patternfly/react-templates
+
+# [6.0.0-prerelease.11](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-templates@6.0.0-prerelease.10...@patternfly/react-templates@6.0.0-prerelease.11) (2024-09-17)
+
+**Note:** Version bump only for package @patternfly/react-templates
+
+# [6.0.0-prerelease.10](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-templates@6.0.0-prerelease.9...@patternfly/react-templates@6.0.0-prerelease.10) (2024-09-16)
+
+**Note:** Version bump only for package @patternfly/react-templates
+
+# [6.0.0-prerelease.9](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-templates@6.0.0-prerelease.8...@patternfly/react-templates@6.0.0-prerelease.9) (2024-09-13)
+
+**Note:** Version bump only for package @patternfly/react-templates
+
+# [6.0.0-prerelease.8](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-templates@6.0.0-prerelease.7...@patternfly/react-templates@6.0.0-prerelease.8) (2024-09-13)
+
+**Note:** Version bump only for package @patternfly/react-templates
+
+# [6.0.0-prerelease.7](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-templates@6.0.0-prerelease.6...@patternfly/react-templates@6.0.0-prerelease.7) (2024-09-12)
+
+**Note:** Version bump only for package @patternfly/react-templates
+
 # [6.0.0-prerelease.6](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-templates@6.0.0-prerelease.5...@patternfly/react-templates@6.0.0-prerelease.6) (2024-09-12)
 
 **Note:** Version bump only for package @patternfly/react-templates